Rename seedu.address.network.* to seedu.address.util.*
