+++ src/test/java/seedu/address/util/NetworkTest.java
@@ -1,15 +1,15 @@
-package seedu.address.network;
+package seedu.address.util;
 
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
-import static seedu.address.network.Network.fetchAsStream;
-import static seedu.address.network.Network.fetchAsString;
 import static seedu.address.testutil.TypicalEntries.FILE_TEST_CONTENTS;
 import static seedu.address.testutil.TypicalEntries.VALID_FILE_LINK;
 import static seedu.address.testutil.TypicalEntries.VALID_HTTPS_LINK;
 import static seedu.address.testutil.TypicalEntries.VALID_HTTP_LINK;
+import static seedu.address.util.Network.fetchAsStream;
+import static seedu.address.util.Network.fetchAsString;
 
 import java.io.IOException;
 import java.io.InputStream;
